Alterações no backend
